+++ public/scripts/validate/validateInfo.js
@@ -45,11 +45,10 @@ async function validate(displayname, username, password, confirmpassword) {
                         fixedDisplayname = fixedDisplayname.trimRight();
 
                         //console.log(fixedDisplayname)
-
-                        const body = { "username": username.toLowerCase(), "password": password, "displayname": fixedDisplayname };
+                        const infoHeader = { "authorization": "Basic " + window.btoa(`${username.toLowerCase()}:${password}:${fixedDisplayname}`) };
                         const url = `/access`;
 
-                        const resp = await callServerAPI(body, url);
+                        const resp = await callServerAPIPost(infoHeader, url);
 
                         if (resp) {
                             message = resp;//"godkjent";
@@ -91,15 +90,15 @@ async function login(username, password, rmbrMe) {
 
         if (username.length >= minCharLength && username.length <= maxCharLength) {
 
-            const body = { "authorization": "Basic " + window.btoa(`${username}:${password}`) };
+            const infoHeader = { "authorization": "Basic " + window.btoa(`${username}:${password}`) };
 
             const url = `/authenticate`;
 
-            const resp = await callServerAPI(body, url);
+            const resp = await callServerAPIPost(infoHeader, url);
 
             if (resp) {
 
-                if (resp.authToken) {
+                if (resp.authToken && resp.user && resp.user.details && resp.user.settings) {
 
                     localStorage.clear();
                     sessionStorage.clear();
@@ -107,14 +106,14 @@ async function login(username, password, rmbrMe) {
                     if (rmbrMe === true) {
                         localStorage.setItem("authToken", resp.authToken);
                         localStorage.setItem("user", JSON.stringify(resp.user));
+                        localStorage.setItem("userDetails", JSON.stringify(resp.user.details));
+                        localStorage.setItem("userSettings", JSON.stringify(resp.user.settings));
                     } else {
                         sessionStorage.setItem("authToken", resp.authToken);
                         sessionStorage.setItem("user", JSON.stringify(resp.user));
+                        sessionStorage.setItem("userDetails", JSON.stringify(resp.user.details));
+                        sessionStorage.setItem("userSettings", JSON.stringify(resp.user.settings));
                     }
-
-                    const today = new Date();
-                    const updatedTxt = today.toLocaleDateString() || true;
-                    sessionStorage.setItem("updated", updatedTxt);
 
                     message = "Login successful";
                 } else {
